Replace do-while loop with for loop
The do-while ends after N iterations, where x*2^16 + N*dx >= x2*2^16. This implies N >= L. For this reason, L is the total number of iterations and I've renamed it accordingly. I suspect that this simplification is also getting rid of a bug where Dot_3x3 would receive y = 2^16-1 as an argument.