Restore do-while loop with safeguards
This is done to preserve the nice-looking overdraw of the old algorithm, without incurring in any nasty crashes.

If we absolutely need to optimize this loop for performance, we could break it down into a for(i = 0; i<iterations; i++) followed by a while loop for safe overdrawing. But there's no need for now.